requirements.txt
torch~=1.13.1
torchvision~=0.14.1
numpy~=1.22.3
Pillow~=9.1.1